% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot2pval.R
\name{boot2pval}
\alias{boot2pval}
\title{My function of bootstraped p values}
\usage{
boot2pval(x1, x2, conf.level = 0.95, iter = 3000, mudiff = 0, test = "two")
}
\arguments{
\item{x1}{a sample vector}

\item{x2}{a sample vector to be subtracted from x1}

\item{conf.level}{the confidence level}

\item{iter}{the number of iterations for bootstrap sampling}

\item{mudiff}{the null value}

\item{test}{the type of t test to perform}
}
\value{
graphical explanations of the t test as well as key parameters
}
\description{
A function to conduct t tests using bootstrap sampling methods as well as return useful graphs.
}
\examples{
boot2pval(x1 = y, x2 = x, mudiff = 0)
}
