% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxlikg2.R
\name{maxlikg2}
\alias{maxlikg2}
\title{My function to calculate maximum likelihood of two variables}
\usage{
maxlikg2(theta1, theta2, lfun = "logbinpois", ...)
}
\arguments{
\item{theta1}{a sequence of values that the first parameter could be}

\item{theta2}{a sequence of values that the second parameter could be}

\item{lfun}{desired function to apply to the two values}

\item{...}{graphical parameters}
}
\value{
a plot with a point estimate showing the two estimated values
}
\description{
Produces a plot to derive the maximum likelihood in the event multivariate analysis is needed
}
\details{
This is part of the lab 10 function requirement
}
\examples{
maxlikg2(seq(0,1,length = 1000), seq(0,10, length = 1000), lfun = "logbinpois")
}
