% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{My function of binomial sample simulation}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter, }{a number of iterations}

\item{n, }{the sample size}

\item{p, }{the probability of "success"}
}
\value{
a barplot showing the simulated distribution
}
\description{
Produces barplots which are binomially distributed
}
\details{
This is a part of the lab 5 function requirement
}
\examples{
x = 1:57 ; zvals(x)
}
