% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{My function to create bootstrap samples}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter, }{a number of iterations}

\item{x, }{a sample variable}

\item{fun, }{a fucntion to apply to the sample}

\item{alpha, }{confidence}

\item{cx, }{size of abrs}

\item{...}{}
}
\value{
a barplot displaying the test statistic and the confidence interval
}
\description{
Produces barplots and sample statistics based on parameters given
}
\details{
This is a part of the lab 9 function requirement
}
\examples{
myboot2(x = c(1:10), fun = "median", alpha = 0.2)
}
