% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{My function to create a 95% confidence interval}
\usage{
myci(x)
}
\arguments{
\item{x}{a vector representation of a sample}
}
\value{
a confidence interval
}
\description{
This is part of the lab 11 function requirement
}
\examples{
x = rnorm(30, mean =10, sd = 12); myci(x)
}
