% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{My function of poisson central limit theorem graphs}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{...}{}
}
\value{
a set of graphs relating to poisson distribution
}
\description{
Produces a set of graphs for given poisson distribution parameters
}
\details{
This is a part of the lab 8 requirement
}
\examples{
mycltp(n = 10, iter = 10000, lambda = 5)
}
