% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{My function of creating a normal distribution plots}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu, }{the mean}

\item{sigma, }{the standard deviation}

\item{a, }{a sample point}
}
\value{
a probability and a density plot
}
\description{
Produces cumulative normal density plots and its area.
}
\details{
This is a part of the lab 6 function requirement
}
\examples{
mean = 0, sd = 1, a = 2; myncurve(0,1,2)
}
