% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zvals.R
\name{zvals}
\alias{zvals}
\title{My function of z scores}
\usage{
zvals(x)
}
\arguments{
\item{x}{a dataset}
}
\value{
a list of vectors
}
\description{
Produces a vector of z scores for a given data set
}
\details{
This is a part of the lab 4 requirement, but it is from lab 2
}
\examples{
x = 1:58 ; zvals(x)
}
